feat: update to 2020 VC context and security-v3-unstable context
This updates to a working version of v3-unstable context for security-vocab and uses an updated 2020
VC context that doesn't conflict with the V3-unstable context.

BREAKING CHANGE: The context URLs change in the fixture that causes a breaking change to other
implementations that do not support these URLs locally and strictly resolve local contexts only.

follow up for issue #68