fix: addresses bug with blank nodes that was breaking nested reveals (#96)
the blank nodes were only being transformed in the node identifier portion of the nQuad. This caused
issues in how the nQuads would be converted back from RDF to JSON which in-advertently broke the
ability for a proof to be derived from a nested structure. This fix makes it so that all blank node
identifiers are transformed properly rather than just the node identifiers (first part of the nQuad).

fixes #91